fifth commit

+++ geometricMeanCalculator.c
@@ -2,14 +2,14 @@
 #include <math.h>
 
 // geometric mean calculator function
-float geometricMeanCalculator(int count, float *arr)
+double geometricMeanCalculator(int count, double *arr)
 {
-    float product = 1, geometricMean;
+    double product = 1, geometricMean;
     for (int i = 0; i < count; i++)
     {
         product *= arr[i];
     }
-    geometricMean = pow(product, (1/(double)count));
+    geometricMean = pow(product, (1 / (double)count));
     return geometricMean;
 }
 
@@ -18,13 +18,13 @@ int main()
     int count;
     printf("How many numbers do you wanna input? : ");
     scanf("%d", &count);
-    float arr[count - 1];
+    double arr[count - 1];
 
     // taking inputs
     for (int i = 0; i < count; i++)
     {
         printf("Enter value: ");
-        if (scanf("%f", &arr[i]) == 0)
+        if (scanf("%lf", &arr[i]) == 0)
         {
             printf("Wrong input. program terminated");
             return -1;
@@ -32,11 +32,11 @@ int main()
     }
 
     // calculating geometric mean
-    float geometricMean;
+    double geometricMean;
     geometricMean = geometricMeanCalculator(count, arr);
 
     // printing the geometric mean
-    printf("\nThe geometric mean of the given data is: %f", geometricMean);
+    printf("\nThe geometric mean of the given data is: %lf", geometricMean);
 
     return 0;
 }
