+++ medianCalculator.c
@@ -1,14 +1,14 @@
 #include <stdio.h>
 
 // swap function
-void swap(float *first, float *second)
+void swap(double *first, double *second)
 {
-    float temp;
+    double temp;
     temp = *first, *first = *second, *second = temp;
 }
 
 // sorting given array
-void sortArr(int count, float *arr)
+void sortArr(int count, double *arr)
 {
     for (int i = 0; i < count; i++)
     {
@@ -23,7 +23,7 @@ void sortArr(int count, float *arr)
 }
 
 // median calculator function
-float medianCalculator(int count, float *sorted_arr)
+double medianCalculator(int count, double *sorted_arr)
 {
     if (count % 2 == 0)
         return (sorted_arr[(count - 1) / 2] + sorted_arr[count / 2]) / 2.0;
@@ -35,13 +35,13 @@ int main()
     int count;
     printf("How many numbers do you wanna input? : ");
     scanf("%d", &count);
-    float arr[count - 1];
+    double arr[count - 1];
 
     // taking inputs
     for (int i = 0; i < count; i++)
     {
         printf("Enter value: ");
-        if (scanf("%f", &arr[i]) == 0)
+        if (scanf("%lf", &arr[i]) == 0)
         {
             printf("Wrong input. program terminated");
             return -1;
@@ -52,11 +52,11 @@ int main()
     sortArr(count, arr);
 
     // calculating median
-    float median;
+    double median;
     median = medianCalculator(count, arr);
 
     // printing the median
-    printf("\nThe median of the given data is: %f", median);
+    printf("\nThe median of the given data is: %lf", median);
 
     return 0;
 }
